+++ WooCommerce/WooCommerceTests/ViewRelated/Orders/Order Creation/ProductRowViewModelTests.swift
@@ -176,7 +176,7 @@ final class ProductRowViewModelTests: XCTestCase {
 
         // When
         let viewModel = ProductRowViewModel(product: product, canChangeQuantity: false, currencyFormatter: currencyFormatter)
-        viewModel.incrementQuantity()
+        viewModel.stepperViewModel.incrementQuantity()
 
         // Then
         let expectedPriceLabel = "$5.00"
@@ -350,174 +350,6 @@ final class ProductRowViewModelTests: XCTestCase {
         // Then
         XCTAssertTrue(viewModel.orderProductDetailsLabel.contains(ProductType.bundle.description), "Label should contain product type (Bundle)")
         XCTAssertTrue(viewModel.orderProductDetailsLabel.contains(stockStatus.description), "Label should contain stock status")
-    }
-
-    func test_increment_and_decrement_quantity_have_step_value_of_one() {
-        // Given
-        let product = Product.fake()
-        let viewModel = ProductRowViewModel(product: product, canChangeQuantity: true)
-
-        // When & Then
-        viewModel.incrementQuantity()
-        XCTAssertEqual(viewModel.quantity, 2)
-
-        // When & Then
-        viewModel.decrementQuantity()
-        XCTAssertEqual(viewModel.quantity, 1)
-    }
-
-    func test_quantity_has_minimum_value_of_one() {
-        // Given
-        let product = Product.fake()
-        let viewModel = ProductRowViewModel(product: product, canChangeQuantity: true)
-        XCTAssertEqual(viewModel.quantity, 1)
-
-        // When
-        viewModel.decrementQuantity()
-
-        // Then
-        XCTAssertEqual(viewModel.quantity, 1)
-    }
-
-    func test_cannot_decrement_quantity_below_zero() {
-        // Given
-        let product = Product.fake()
-        let viewModel = ProductRowViewModel(product: product, quantity: 0, canChangeQuantity: true)
-        XCTAssertEqual(viewModel.quantity, 0)
-
-        // When
-        viewModel.decrementQuantity()
-
-        // Then
-        XCTAssertEqual(viewModel.quantity, 0)
-        XCTAssertTrue(viewModel.shouldDisableQuantityDecrementer, "Quantity decrementer is not disabled")
-    }
-
-    func test_decrement_quantity_at_minimum_quantity_removes_product() {
-        // Given
-        let product = Product.fake()
-        var productRemoved = false
-        let viewModel = ProductRowViewModel(product: product, canChangeQuantity: true, removeProductIntent: { productRemoved = true })
-
-        // When
-        viewModel.decrementQuantity()
-
-        // Then
-        XCTAssertTrue(productRemoved)
-    }
-
-    func test_cannot_decrement_quantity_below_minimumQuantity() {
-        // Given
-        let viewModel = ProductRowViewModel(productOrVariationID: 1,
-                                            name: "",
-                                            sku: nil,
-                                            price: nil,
-                                            stockStatusKey: "",
-                                            stockQuantity: nil,
-                                            manageStock: false,
-                                            quantity: 3,
-                                            minimumQuantity: 3,
-                                            maximumQuantity: nil,
-                                            canChangeQuantity: true,
-                                            imageURL: nil,
-                                            hasParentProduct: false,
-                                            isConfigurable: false)
-        XCTAssertEqual(viewModel.quantity, 3)
-
-        // When
-        viewModel.decrementQuantity()
-
-        // Then
-        XCTAssertEqual(viewModel.quantity, 3)
-    }
-
-    func test_cannot_increment_quantity_beyond_maximumQuantity() {
-        // Given
-        let viewModel = ProductRowViewModel(productOrVariationID: 1,
-                                            name: "",
-                                            sku: nil,
-                                            price: nil,
-                                            stockStatusKey: "",
-                                            stockQuantity: nil,
-                                            manageStock: false,
-                                            quantity: 6,
-                                            minimumQuantity: 4,
-                                            maximumQuantity: 6,
-                                            canChangeQuantity: true,
-                                            imageURL: nil,
-                                            hasParentProduct: false,
-                                            isConfigurable: false)
-        XCTAssertEqual(viewModel.quantity, 6)
-
-        // When
-        viewModel.incrementQuantity()
-
-        // Then
-        XCTAssertEqual(viewModel.quantity, 6)
-    }
-
-    func test_quantity_decrementer_disabled_at_minimum_quantity_when_removeProductIntent_is_nil() {
-        // Given
-        let viewModel = ProductRowViewModel(productOrVariationID: 1,
-                                            name: "",
-                                            sku: nil,
-                                            price: nil,
-                                            stockStatusKey: "",
-                                            stockQuantity: nil,
-                                            manageStock: false,
-                                            quantity: 3,
-                                            minimumQuantity: 3,
-                                            maximumQuantity: nil,
-                                            canChangeQuantity: true,
-                                            imageURL: nil,
-                                            hasParentProduct: false,
-                                            isConfigurable: false)
-
-        // Then
-        XCTAssertTrue(viewModel.shouldDisableQuantityDecrementer)
-    }
-
-    func test_quantity_decrementer_not_disabled_at_minimum_quantity_when_removeProductIntent_is_not_nil() {
-        // Given
-        let viewModel = ProductRowViewModel(productOrVariationID: 1,
-                                            name: "",
-                                            sku: nil,
-                                            price: nil,
-                                            stockStatusKey: "",
-                                            stockQuantity: nil,
-                                            manageStock: false,
-                                            quantity: 3,
-                                            minimumQuantity: 3,
-                                            maximumQuantity: nil,
-                                            canChangeQuantity: true,
-                                            imageURL: nil,
-                                            hasParentProduct: false,
-                                            isConfigurable: false,
-                                            removeProductIntent: {})
-
-        // Then
-        XCTAssertFalse(viewModel.shouldDisableQuantityDecrementer)
-    }
-
-    func test_quantity_incrementer_disabled_at_maximum_quantity() {
-        // Given
-        let viewModel = ProductRowViewModel(productOrVariationID: 1,
-                                            name: "",
-                                            sku: nil,
-                                            price: nil,
-                                            stockStatusKey: "",
-                                            stockQuantity: nil,
-                                            manageStock: false,
-                                            quantity: 6,
-                                            minimumQuantity: 4,
-                                            maximumQuantity: 6,
-                                            canChangeQuantity: true,
-                                            imageURL: nil,
-                                            hasParentProduct: false,
-                                            isConfigurable: false)
-
-        // Then
-        XCTAssertTrue(viewModel.shouldDisableQuantityIncrementer)
     }
 
     func test_productRow_when_add_discount_button_is_tapped_then_orderProductDiscountAddButtonTapped_is_tracked() {
